pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gimme-your-words"
version = "0.2.0"
description = "Download articles from the web for personal use"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "playwright>=1.40.0",
    "markdownify>=0.11.6",
    "click>=8.1.0",
    "playwright-stealth>=1.0.6",
    "pyyaml>=6.0",
]

[project.scripts]
gimme-your-words = "gimme_your_words.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["gimme_your_words"]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
]
